add 게시글 상세보기 -> 댓글 list write form 추가

+++ communityJh/src/main/java/kr/co/communityJh/repository/BoardRepository.java
@@ -11,7 +11,7 @@ import kr.co.communityJh.entity.Board;
  * @author jhlee
  * Board JpaRepository
  */
-public interface BoardRepository extends JpaRepository<Board, Integer>, QuerydslPredicateExecutor<Board>{
+public interface BoardRepository extends JpaRepository<Board, Long>, QuerydslPredicateExecutor<Board>{
 	
 	Page<Board> findByTitleContainingOrBodyContaining(String title, String body, Pageable pageable);
 //	Page<Board> findByNicknameContaining(String nickName, Pageable pageable);
